Fix docs for importing vpc_subnet_v1 (#1081)
Fix docs for importing vpc_subnet_v1

Summary of the Pull Request
This PR fixes an issue in the documentation for importing vpc_subnet_v1 resources.
It was mentioned that the subnet id should be used for importing the resource, which didn't work (failed silently). Checking via terraform state list didn't contain the imported resource.
The only hint I had was this warning shown with TF_LOG=TRACE:
2021-05-24T15:35:08.897+0200 [WARN]  Provider "registry.terraform.io/opentelekomcloud/opentelekomcloud" produced an unexpected new value for opentelekomcloud_vpc_subnet_v1.subnet["<RESOURCE INSTANCE NAME>"] during refresh.
      - Root resource was present, but now absent

So I tried using the network id of the subnet for the import, which worked.
PR Checklist

 Refers to: #xxx
 Tests added/passed.
 Documentation updated.
 Schema updated.

Reviewed-by: Rodion Gyrbu <fpsoff@outlook.com>
Reviewed-by: None <None>
Reviewed-by: Anton Kachurin <katchuring@gmail.com>