[RDS]: PostgreSql 13/14 workaround fix to enable `parameters` (#2068)
[RDS]: PostgreSql 13/14 workaround fix to enable `parameters`

Summary of the Pull Request
PostgreSql 13/14 have issues which leads to an error when applying parameters.
This PR is a workaround meant to fix this issue.
PR Checklist

 Refers to: #2051
 Tests added/passed.
 Documentation updated.
 Schema updated.
 Release notes added.

Acceptance Steps Performed
=== RUN   TestAccRdsInstanceV3RestoreBackup
--- PASS: TestAccRdsInstanceV3RestoreBackup (869.82s)
PASS

Process finished with the exit code 0

=== RUN   TestAccRdsInstanceV3ElasticIP
--- PASS: TestAccRdsInstanceV3ElasticIP (454.12s)
PASS

Process finished with the exit code 0

=== RUN   TestAccRdsInstanceV3Basic
--- PASS: TestAccRdsInstanceV3Basic (715.05s)
PASS

Process finished with the exit code 0

=== RUN   TestAccRdsPostgre13V3Basic
--- PASS: TestAccRdsPostgre13V3Basic (446.41s)
PASS

Process finished with the exit code 0

Reviewed-by: Aloento
Reviewed-by: Anton Sidelnikov