Least amount of KMS roles for server_side_encryption (#2371)
Least amount of KMS roles for server_side_encryption

Summary of the Pull Request
Hello there,
It took me some time to figure out what is the least amount of roles needed to use an encrypted bucket with an OTC user. Thought I would share it with the community. If there is a better place for this, please point me to that direction.
PR Checklist

 Refers to: #xxx
 Tests added/passed.
 Documentation updated.
 Schema updated.
 Release notes added.

Acceptance Steps Performed
=== RUN   TestAccSomethingV0_basic
--- PASS: TestAccSomethingV0_basic (101.71s)
=== RUN   TestAccSomethingV0_timeout
--- PASS: TestAccSomethingV0_timeout (128.67s)
PASS

Process finished with exit code 0

Reviewed-by: Anton Sidelnikov
Reviewed-by: Artem Lifshits