Update addon-templates-v1.19.8.md (#2418)
Update addon-templates-v1.19.8.md

Summary of the Pull Request
hello there,
this was causing some confusion for one of the OTC customers. What is happening if you have is_driver_from_nvidia = false (which is default on the API) then in the CCE deployment of GPU beta your URL will be prefixed with obs:// which will break the public URL you are trying to use.
PR Checklist

 Refers to: #xxx
 Tests added/passed.
 Documentation updated.
 Schema updated.
 Release notes added.

Acceptance Steps Performed
=== RUN   TestAccSomethingV0_basic
--- PASS: TestAccSomethingV0_basic (101.71s)
=== RUN   TestAccSomethingV0_timeout
--- PASS: TestAccSomethingV0_timeout (128.67s)
PASS

Process finished with exit code 0

Reviewed-by: Anton Sidelnikov
Reviewed-by: Artem Lifshits